!735 mome 新算子支持
* bugfix: mome metadata mtp 适配和cp修改
* mome layer.gate float 去除
* 修复build_for_cudagraph_capture的构建导致的入图精度问题
* Merge branch 'omninpu_pg2_perf' of https://gitee.com/omniai/omni-npu i…
* 解决冲突
* Merge branch 'omninpu_pg2_perf' of https://gitee.com/omniai/omni-npu i…
* mome update block table 修复
* bugfix: page size compute
* assert 补充
* mome 新算子支持

+++ src/omni_npu/layers/fused_moe/layer.py
@@ -97,7 +97,7 @@ class NPUUnquantizedFusedMoEMethod(UnquantizedFusedMoEMethod, NPUFusedMoEMethodB
             x_slice = x_slice[start:end]
 
         if layer.gate is not None:
-            router_logits, _ = layer.gate(x_slice.float())
+            router_logits, _ = layer.gate(x_slice)
         else:
             assert router_logits is not None, "Expected gate or router_logits must be provided."
             if is_need_slice:
